fixup! [tools/RA-TLS] Don't use insecure unhex and rework secret-prov examples

+++ tools/sgx/ra-tls/secret_prov_attest.c
@@ -413,7 +413,7 @@ __attribute__((constructor)) static void secret_provision_constructor(void) {
 
             char path_buf[256];
             ret = snprintf(path_buf, sizeof(path_buf), "/dev/attestation/keys/%s", key_name);
-            if (ret < 0 || (size_t)ret >= (ssize_t)sizeof(path_buf)) {
+            if (ret < 0 || (size_t)ret >= sizeof(path_buf)) {
                 ERROR("Provisioned key name '%s' is too long\n", key_name);
                 exit(1);
             }
